lag_hr_cnt can be a float value

+++ xerparser/model/classes/taskpred.py
@@ -28,7 +28,7 @@ class TaskPred:
         self.proj_id = int(params.get('proj_id').strip()) if params.get('proj_id') else None
         self.pred_proj_id = int(params.get('proj_id').strip()) if params.get('pred_proj_id') else None
         self.pred_type = params.get('pred_type').strip() if params.get('pred_type') else None
-        self.lag_hr_cnt = int(params.get('lag_hr_cnt').strip()) if params.get('lag_hr_cnt') else None
+        self.lag_hr_cnt = float(params.get('lag_hr_cnt').strip()) if params.get('lag_hr_cnt') else None
         self.float_path = params.get('float_path').strip() if params.get('float_path') else None
         self.aref = params.get('aref').strip() if params.get('aref') else None
         self.arls = params.get('arls').strip() if params.get('arls') else None
